requirements.txt
amqp==2.3.2
astroid==2.0.4
billiard==3.5.0.4
Django==2.1.1
django-compat==1.0.15
djangorestframework==3.8.2
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.5
pylint==2.1.1
pytz==2018.5
six==1.11.0
vine==1.1.4
wrapt==1.10.11
